refactor: presentation 패키지 정리 (#52)
