+++ app/src/main/java/com/hyeeyoung/wishboard/config/navigation/navgraph/ItemDetailNavGraph.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import com.hyeeyoung.wishboard.config.navigation.screen.MainScreen
-import com.hyeeyoung.wishboard.presentation.wish.WishItemDetailScreen
+import com.hyeeyoung.wishboard.presentation.wish.screen.WishItemDetailScreen
 
 fun NavGraphBuilder.itemDetailNavGraph(navController: NavHostController) = composable(
     route = MainScreen.WishItemDetail.routeWithArg,
